requirements.txt
APScheduler==3.6.3
SQLAlchemy==1.4.7
certifi==2020.12.5
greenlet==1.0.0
mysqlclient==2.0.3
pip==21.0.1
python-dotenv==0.17.0
python-telegram-bot==13.4.1
pytz==2021.1
secure-smtplib==0.1.1
setuptools==56.0.0
six==1.15.0
tornado==6.1
tzlocal==2.1